requirements/common.txt
Django==1.9.4
djangorestframework==3.3.3
psycopg2==2.6.1
djoser==0.4.3
django-allauth==0.25.2
django-cors-headers==1.1.0
django-filter==0.12.0
django-crispy-forms==1.6.0
django-formtools==1.0
django-countries==3.4.1
django-leaflet==0.18.0
jsonschema==2.5.1
rfc3987==1.3.5
drfdocs==0.0.9
django-tutelary==0.1.9
django-audit-log==0.7.0
simplejson==3.8.1
django-widget-tweaks==1.4.1
